+++ tutorial08/main.cpp
@@ -33,7 +33,7 @@ static void timerFunc(int)
 {
     double dt = 1.0/double(frameRate);
 
-    unsigned int millis = (unsigned int)(dt * 1000);
+    unsigned int millis = static_cast<unsigned int>(dt * 1000);
     glutTimerFunc(millis, timerFunc, 1);
 
     Basic::Timer::updateTimers(dt);
@@ -104,8 +104,8 @@ int main(int argc, char* argv[])
   sys->createWindow();
 
   // set timer
-  double dt = 1.0/double(frameRate);
-  unsigned int millis = (unsigned int)(dt * 1000);
+  double dt = 1.0/static_cast<double>(frameRate);
+  unsigned int millis = static_cast<unsigned int>(dt * 1000);
   glutTimerFunc(millis, timerFunc, 1);
 
   // main loop
